Update!: implement minimax in backpropagation instead of exploration.
Instead of inverting the score when selecting the best children in
exploration, this version backpropagates a game outcome instead of a
single score up the tree, and each node gets assigned a win or loss
according to which player is supposed to make the play in the node and
who the winner actually is.

I feel like this should be basically the same thing as inverting the
score when exploring. I don't understand why but this version just
seems to work better than 74e6127066dc.