Fix!: the bug in 724691ccc582.
Although this probably is also buggy. It's sending an
`Rc<RefCell<usize>>` to an async task. I don't think that's supposed to
work? Or at least it's maybe a potential crash point.

More of a workaround. Keep an index of the game, increment it whenever
the game restarts; discard the result from AI if the index after the
calculation doesn't match the index before it started.

Probably should look into properly worker task cancellation.